Added readonly editor for style congruency
