+++ components/editor/editor/RenderPost.tsx
@@ -1,25 +1,22 @@
-import React, { useMemo } from "react";
-import DOMPurify, { Config } from "dompurify";
-import { generateHTML } from "@tiptap/core";
-import { TiptapExtensions } from "./extensions";
+import React from "react";
+import { CustomCodeBlockReadOnly, TiptapExtensions } from "./extensions";
 import "highlight.js/styles/monokai-sublime.css";
+import { EditorContent, useEditor } from "@tiptap/react";
 
 interface RenderPostProps {
   json: string;
 }
 
-const config: Config = {
-  ADD_TAGS: ["iframe"],
-  ADD_ATTR: ["allowfullscreen", "target"],
-};
-
 const RenderPost = ({ json }: RenderPostProps) => {
-  const sanitizedHTML = useMemo(() => {
-    const rawHTML = generateHTML(JSON.parse(json), [...TiptapExtensions]);
-    return DOMPurify.sanitize(rawHTML, config) as string;
-  }, [json]);
+  const content = JSON.parse(json);
 
-  return <div dangerouslySetInnerHTML={{ __html: sanitizedHTML }} />;
+  const editor = useEditor({
+    editable: false,
+    extensions: [...TiptapExtensions, CustomCodeBlockReadOnly],
+    content,
+  });
+
+  return <EditorContent editor={editor} />;
 };
 
 export default RenderPost;
