+++ components/editor/editor/components/CodeBlock/CodeBlock.module.css
@@ -1,0 +1,4 @@
+.code-block div {
+  background-color: black;
+  color: white;
+}
